Now allowing range rings with small distances below one NM/km/mi. Improved range rings distance validation in options dialog for easier input. Added distance to navaid range ring label. Now either disabling menu items for navaids without range or showing normal range rings in case of modifier clicks to avoid zero range circles.

+++ src/common/htmlinfobuilder.cpp
@@ -2059,7 +2059,7 @@ void HtmlInfoBuilder::rangeMarkerText(const RangeMarker& marker, atools::util::H
   else
   {
     QStringList distStr;
-    for(int dist : marker.ranges)
+    for(float dist : marker.ranges)
       distStr.append(Unit::distNm(dist, false));
 
     html.brText((marker.ranges.size() > 1 ? tr("Distances: %1 %2") : tr("Distance: %1 %2")).
