+++ src/common/maptypes.h
@@ -1012,7 +1012,7 @@ struct RangeMarker
   }
 
   QString text; /* Text to display like VOR name and frequency */
-  QVector<int> ranges; /* Range ring list (nm) */
+  QVector<float> ranges; /* Range ring list (nm) */
   MapTypes type; /* VOR, NDB, AIRPORT, etc. */
 };
